Moved containers to new directory
