+++ engine/benchmark/CMakeLists.txt
@@ -31,3 +31,4 @@ add_engine_benchmark(graphics)
 add_engine_benchmark(ecs)
 add_engine_benchmark(async)
 add_engine_benchmark(random)
+add_engine_benchmark(containers)
