+++ engine/benchmark/containers/type_map_bench.cpp
@@ -1,4 +1,4 @@
-#include "genebits/engine/util/type_map.h"
+#include "genebits/engine/containers/type_map.h"
 
 #include <map>
 #include <unordered_map>
