+++ engine/test/CMakeLists.txt
@@ -34,3 +34,4 @@ add_engine_test(ecs)
 add_engine_test(async)
 add_engine_test(os)
 add_engine_test(random)
+add_engine_test(containers)
