+++ engine/test/containers/fast_vector_tests.cpp
@@ -1,4 +1,4 @@
-#include "genebits/engine/util/fast_vector.h"
+#include "genebits/engine/containers/fast_vector.h"
 
 #include <gtest/gtest.h>
