+++ engine/test/containers/type_map_tests.cpp
@@ -1,4 +1,4 @@
-#include "genebits/engine/util/type_map.h"
+#include "genebits/engine/containers/type_map.h"
 
 #include <gtest/gtest.h>
